moved blog post query outside of layout file
