+++ next.config.js
@@ -1,9 +1,13 @@
-module.exports = {
+// @ts-check
+
+/**
+ * @type {import('next').NextConfig}
+ */
+const nextConfig = {
   reactStrictMode: true,
   images: {
     domains: ["images.ctfassets.net"],
   },
-  env: {
-    BASE_URL: process.env.BASE_URL,
-  },
 };
+
+module.exports = nextConfig;
