+++ package.json
@@ -91,7 +91,7 @@
     "postcss": "^8.4.5",
     "prettier": "^2.6.2",
     "tailwindcss": "^3.0.12",
-    "typescript": "^4.6.4",
+    "typescript": "^5.1.6",
     "vitest": "^0.34.1"
   },
   "readme": "https://github.com/dtran421/portfolio#readme",
